+++ app/src/main/java/br/alexandregpereira/jerry/app/animation/ExpandFadingSpringAnimationActivity.kt
@@ -5,10 +5,10 @@ import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import br.alexandregpereira.jerry.app.R
-import br.alexandregpereira.jerry.expandable.goneCollapseHeightFadeOut
-import br.alexandregpereira.jerry.expandable.goneCollapseWidthFadeOut
-import br.alexandregpereira.jerry.expandable.visibleExpandHeightFadeIn
-import br.alexandregpereira.jerry.expandable.visibleExpandWidthFadeIn
+import br.alexandregpereira.jerry.animation.expandable.goneCollapseHeightFadeOut
+import br.alexandregpereira.jerry.animation.expandable.goneCollapseWidthFadeOut
+import br.alexandregpereira.jerry.animation.expandable.visibleExpandHeightFadeIn
+import br.alexandregpereira.jerry.animation.expandable.visibleExpandWidthFadeIn
 import kotlinx.android.synthetic.main.activity_expand_fading_animation.*
 
 class ExpandFadingSpringAnimationActivity : AppCompatActivity() {
