+++ jerry/src/main/java/br/alexandregpereira/jerry/animation/TransalationSpringAnimation.kt
@@ -1,8 +1,11 @@
-package br.alexandregpereira.jerry
+package br.alexandregpereira.jerry.animation
 
 import android.view.View
 import androidx.dynamicanimation.animation.DynamicAnimation
-import androidx.dynamicanimation.animation.SpringForce
+import br.alexandregpereira.jerry.JerryAnimation
+import br.alexandregpereira.jerry.JerryAnimationSet
+import br.alexandregpereira.jerry.SpringAnimationPropertyKey
+import br.alexandregpereira.jerry.spring
 
 fun JerryAnimation.translationXSpring(
     targetValue: Float
